refactor!: footer removal
GetFormState() should now be used as the footer no longer exists. GetFooterStatusValue() has been marked as obsolete and calls GetFormState() internally so that this is a non-breaking change.

BREAKING CHANGE: XPath selectors for footer related elements have been removed. GetFooterMessageValue() has been removed (unaware of a replacement for this so unable to mark obsolete and redirect to new method).