refactor!: new relevance search
Updates to reflect changes in the UI for Relevance Search in wave 2.

Other than changing the search type (which is no longer possible) these should not be breaking changes and existing tests should continue to work.

BREAKING CHANGE: Removed GlobalSearch.ChangeSearchType as changing the search type is now an admin activity on the environment and users can no longer switch between the two.