feat: 2021 release wave 2 (#1206)
* fix: broken app switching since wave 2 release

* fix: wave 2 xpath for uci app menu button

* fix: backwards compatibility for app-switching

* fix: errors clicking subgrid commands

Since a recent update we've been getting false negatives claiming that subgrid command buttons don't exist. It appears this is actually due to the fact that the 'aria-label' selector used before now includes both the command title and description.

* refactor: rename variables

* fix: switching subgrid view

* fix: open subgrid record

* refactor: revert unintended changes

* style: added linebreak

* Merge branch 'develop' into bugfix/open-subgrid-record

* refactor: line endings

* refactor!: footer removal

GetFormState() should now be used as the footer no longer exists. GetFooterStatusValue() has been marked as obsolete and calls GetFormState() internally so that this is a non-breaking change.

BREAKING CHANGE: XPath selectors for footer related elements have been removed. GetFooterMessageValue() has been removed (unaware of a replacement for this so unable to mark obsolete and redirect to new method).

* refactor!: new relevance search

Updates to reflect changes in the UI for Relevance Search in wave 2.

Other than changing the search type (which is no longer possible) these should not be breaking changes and existing tests should continue to work.

BREAKING CHANGE: Removed GlobalSearch.ChangeSearchType as changing the search type is now an admin activity on the environment and users can no longer switch between the two.

* refactor: new advanced lookup functionality

There is no longer a 'Switch view' button on a lookup results dialog. This commit is a non-breaking change for existing tests that will open 'Advanced lookup' and switch views from there. All Lookup methods have been updated to work from the classic lookup dialog and the advanced lookup dialog.

* refactor: new read-only grids and subgrids

Non-breaking changes to update selectors and logic to work with the new read-only grids and subgrids.

* refactor: footer removal (save)

Updates the selector to use the save button on the ribbon rather than the removed save button on the footer. If this causes problems we may need to use keyboard shortcuts (i.e. CTRL + S).

* fix: improper record set navigator waits

Some failures were observed due to use of WaitForPageToLoad rather than WaitForTransaction. In addition, there was no wait after clicking to open the navigator.

* fix: setting business processs flow boolean fields

These no longer appear as checkboxes but as dropdowns.

* fix: setting date and time fields

The time component of a date and time field isn't shown until focus has been lost from the date field. This meant it was trying to set the time component when it wasn't shown.

* refactor: line endings

* test: fix test not navigating to app

* refactor: grid selectors and get items logic

* fix: entity set navigator closes itself

If you try to open a record in the entity set navigator more than once it will close itself and throw an exception on the second attempt.

* fix: get entity name from DOM for subgrid items

This replaces using Xrm.Page which is deprecated.

* refactor: use correct AppReference XPath

Was using subgrid rather than grid (although currently they are both the same)

* test: fix test broken due to missing view

* test: add comment regarding potential failure

This is now an environment setting rather than a user selected option.

* fix: trying to set time for date only fields

* refactor: line endings

* refactor: line endings

* refactor: line endings