fix: get form state for read only forms (#1220)
* fix: determine form status when form is read-only

Falls back to checking for a state in the header if the record is read-only. A read-only record does not show the record state notification.

* fix: determine form status when form is read-only

Falls back to checking for a state in the header if the record is read-only. A read-only record does not show the record state notification.

* refactor: undo change from other PR